+++ src/lib/validator.js
@@ -1,5 +1,5 @@
 /**
- * Copyright 2018-2020 F5 Networks, Inc.
+ * Copyright 2021 F5 Networks, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -21,6 +21,7 @@ const BigIqSettingsValidator = require('./bigIqSettingsValidator');
 const BusinessLogicValidator = require('./businessLogicValidator');
 const DeviceCertificateValidator = require('./deviceCertificateValidator');
 const LicensePoolValidator = require('./licensePoolValidator');
+const RoutingPrefixListValidator = require('./routingPrefixListValidator');
 const UserValidator = require('./userValidator');
 
 class Validator {
@@ -31,7 +32,8 @@ class Validator {
             new BigIqSettingsValidator(),
             new LicensePoolValidator(),
             new UserValidator(),
-            new DeviceCertificateValidator()
+            new DeviceCertificateValidator(),
+            new RoutingPrefixListValidator()
         ];
     }